Finished the refactoring. The structure should be good now
