+++ WebMarketplace/GlobalUsings.cs
@@ -1,6 +1,6 @@
-global using IWebProduct = SharedClassLibrary.Service.Web.IWebProductService;
+global using IWebProduct = SharedClassLibrary.Service.IProductService;
 global using IWebWaitlist = SharedClassLibrary.Service.IWaitlistService;
 global using IWebNotification = SharedClassLibrary.Service.INotificationService;
-global using WebProduct = SharedClassLibrary.Service.Web.WebProductService;
+global using WebProduct = SharedClassLibrary.Service.ProductService;
 global using WebWaitlist = SharedClassLibrary.Service.WaitlistService;
 global using WebNotification = SharedClassLibrary.Service.NotificationService;
